Change message_ptr from a unique_ptr to a shared_ptr.
* Why?
When a message is passed to each net_con, the only way to support
asynchronous sends is to ensure that the message is not destroyed
until each net_con has finished sending the message. For the
single sake of allowing each net_con to use the message before
it is destroyed, the message_ptr is being changed to a shared_ptr.